Mod Manager - move classes, visibility
