+++ core/src/com/unciv/ui/screens/modmanager/ModUIData.kt
@@ -1,0 +1,82 @@
+package com.unciv.ui.screens.modmanager
+
+import com.badlogic.gdx.scenes.scene2d.ui.TextButton
+import com.unciv.models.metadata.ModCategories
+import com.unciv.models.ruleset.Ruleset
+import com.unciv.models.translations.tr
+import com.unciv.ui.components.extensions.toLabel
+import com.unciv.ui.components.extensions.toTextButton
+import com.unciv.ui.screens.pickerscreens.Github
+
+/** Helper class holds combined mod info for ModManagementScreen, used for both installed and online lists
+ *
+ *  Note it is guaranteed either ruleset or repo are non-null, never both.
+ */
+internal class ModUIData private constructor(
+    val name: String,
+    val description: String,
+    val ruleset: Ruleset?,
+    val repo: Github.Repo?,
+    var y: Float,
+    var height: Float,
+    var button: TextButton
+) {
+    var state = ModStateImages()  // visible only on the 'installed' side - todo?
+
+    constructor(ruleset: Ruleset): this (
+        ruleset.name,
+        ruleset.getSummary().let {
+            "Installed".tr() + (if (it.isEmpty()) "" else ": $it")
+        },
+        ruleset, null, 0f, 0f, getTextButton(ruleset.name, ruleset.modOptions.topics)
+    )
+
+    constructor(repo: Github.Repo, isUpdated: Boolean): this (
+        repo.name,
+        (repo.description ?: "-{No description provided}-".tr()) +
+                "\n" + "[${repo.stargazers_count}]✯".tr(),
+        null, repo, 0f, 0f,
+        getTextButton(repo.name + (if (isUpdated) " - {Updated}" else ""), repo.topics)
+    ) {
+        state.hasUpdate = isUpdated
+    }
+
+
+    fun lastUpdated() = ruleset?.modOptions?.lastUpdated ?: repo?.pushed_at ?: ""
+    fun stargazers() = repo?.stargazers_count ?: 0
+    fun author() = ruleset?.modOptions?.author ?: repo?.owner?.login ?: ""
+
+    fun matchesFilter(filter: ModManagementOptions.Filter): Boolean = when {
+        !matchesCategory(filter) -> false
+        filter.text.isEmpty() -> true
+        name.contains(filter.text, true) -> true
+        // description.contains(filterText, true) -> true // too many surprises as description is different in the two columns
+        author().contains(filter.text, true) -> true
+        else -> false
+    }
+
+    private fun matchesCategory(filter: ModManagementOptions.Filter): Boolean {
+        if (filter.topic == ModCategories.default().topic)
+            return true
+        val modTopics = repo?.topics ?: ruleset?.modOptions?.topics!!
+        return filter.topic in modTopics
+    }
+
+    companion object {
+        private fun getTextButton(nameString: String, topics: List<String>): TextButton {
+            val categories = ArrayList<ModCategories.Category>()
+            for (category in ModCategories) {
+                if (category == ModCategories.default()) continue
+                if (topics.contains(category.topic)) categories += category
+            }
+
+            val button = nameString.toTextButton()
+            val topicString = categories.joinToString { it.label.tr() }
+            if (categories.isNotEmpty()) {
+                button.row()
+                button.add(topicString.toLabel(fontSize = 14))
+            }
+            return button
+        }
+    }
+}
